requirements.txt
langgraph>=0.0.30
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.30

fastapi>=0.100.0
uvicorn>=0.23.0
fastapi-cdn-host>=0.1.4

python-dotenv>=1.0.0
